[REFACT] 키보드 수정
