+++ app/src/main/java/com/preonboarding/customkeyboard/keyboard/KeyboardInterationListener.kt
@@ -1,5 +1,5 @@
 package com.myhome.rpgkeyboard
 
 interface KeyboardInterationListener{
-    fun modechange(mode:Int)
+    fun addView()
 }
