+++ app/src/main/res/drawable-v24/background_itemkeyboard.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/item_keyboard"/>
+</selector>
