+++ app/src/main/res/drawable-v24/item_keyboard.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <corners android:radius="5dp" />
+            <solid android:color="@color/keyboardBoarder" />
+        </shape>
+    </item>
+</layer-list>
