--- app/src/main/res/drawable-v24/key_background.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-            android:state_focused="false"
-            android:state_selected="false"
-            android:state_pressed="false"
-            android:drawable="@drawable/normal"
-    />
-</selector>
