+++ app/src/main/res/layout/item_keyboard.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <Button
-        android:id="@+id/key_button"
+    <TextView
+        android:id="@+id/textview_itemkeyboard"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginLeft="2dp"
-        android:layout_marginTop="4dp"
-        android:layout_marginRight="4dp"
-        android:layout_marginBottom="2dp"
-        android:background="@drawable/key_background"
-        android:duplicateParentState="true"
-        android:longClickable="true" />
+        android:layout_margin="2dp"
+        android:gravity="center"
+        android:textColor="@color/black"
+        android:background="@drawable/background_itemkeyboard"
+        android:textSize="16sp"
+        tools:text="1" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
