+++ app/src/main/res/layout/layout_keyboard.xml
@@ -2,6 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/keyboardBackground"
     android:orientation="vertical">
 
     <LinearLayout
@@ -199,10 +200,10 @@
         android:id="@+id/second_line"
         android:layout_width="match_parent"
         android:layout_height="50dp"
+        android:layout_marginLeft="10dp"
+        android:layout_marginRight="10dp"
         android:layout_weight="1"
-        android:orientation="horizontal"
-        android:paddingLeft="12dp"
-        android:paddingRight="12dp">
+        android:orientation="horizontal">
 
         <include
             android:id="@+id/action_key_a"
@@ -276,7 +277,7 @@
         android:orientation="horizontal">
 
         <include
-            android:id="@+id/action_key_caps"
+            android:id="@+id/action_key_shift"
             layout="@layout/item_keyboard"
             android:layout_width="0dp"
             android:layout_height="match_parent"
@@ -351,14 +352,14 @@
             layout="@layout/item_keyboard"
             android:layout_width="0dp"
             android:layout_height="match_parent"
-            android:layout_weight="2" />
+            android:layout_weight="1.5" />
 
         <include
             android:id="@+id/action_key_mode_change"
             layout="@layout/item_keyboard"
             android:layout_width="0dp"
             android:layout_height="match_parent"
-            android:layout_weight="1" />
+            android:layout_weight="1.5" />
 
         <include
             android:id="@+id/action_key_rest"
